Management - Break useless inheritances: MgtLine.
