+++ models/management/__init__.py
@@ -5,18 +5,12 @@
 # Dependencies
 from . import productivity_day
 
-
 # Management
-
-from . import mgt_line      # Dep !
-
 from . import mgt_order_line
 from . import mgt_family_line
 from . import mgt_subfamily_line
 from . import mgt_day_line
-
 from . import management
-
 
 # Patient
 from . import mgt_patient_line
