+++ models/management/management.py
@@ -9,7 +9,7 @@
 		Define a Strategy - for resolving a problem (business logic).
 
 	Created: 			28 may 2018
-	Last up: 			23 mar 2021
+	Last up: 			24 mar 2021
 """
 from __future__ import print_function
 import collections
