+++ models/management/mgt_day_line.py
@@ -7,7 +7,7 @@
 		day doctor line
 
 	Created: 			15 Jan 2019
-	Last up: 			26 oct 2020
+	Last up: 			24 mar 2021
 """
 
 from __future__ import print_function
@@ -18,14 +18,13 @@ from openerp.addons.openhealth.models.order import ord_vars
 from . import mgt_funcs
 from openerp.addons.openhealth.models.libs import lib
 
-#class DayLine(models.Model):
 class MgtDayLine(models.Model):
 	"""
 	Day Line
 	"""
 	_name = 'openhealth.management.day.line'
 	_order = 'date asc'
-	_inherit = 'openhealth.management.line'
+	#_inherit = 'openhealth.management.line'
 
 
 # ----------------------------------------------------------- Relational -------
@@ -160,3 +159,78 @@ class MgtDayLine(models.Model):
 			#self.avg_amount = np.mean(data_amount)			# Not correct !
 
 	# update_amount
+
+
+# Begin ----------------------------------- Inherited from mgt_line.py ----------------------------
+
+# ----------------------------------------------------------- Interface --------
+	management_id = fields.Many2one(
+			'openhealth.management',
+		)
+
+	doctor_id = fields.Many2one(
+			'openhealth.management.doctor.line',
+			ondelete='cascade',
+		)
+
+# ----------------------------------------------------------- Primitive --------
+	name = fields.Char(
+			'Name',
+		)
+
+	name_sp = fields.Char(
+			'Nombre',
+		)
+
+	meta = fields.Char(
+			'Meta',
+		)
+
+	meta_sp = fields.Char(
+			'Meta',
+		)
+
+	idx = fields.Integer(
+			'Idx',
+		)
+
+	x_count = fields.Integer(
+			'Nr',
+		)
+
+	amount = fields.Float(
+			'Monto',
+			digits=(16, 1),
+		)
+
+	per_amo = fields.Float(
+			'% Monto',
+		)
+
+	per_nr = fields.Float(
+			'% Nr',
+		)
+
+#----------------------------------------------------------- Method ------------
+	@api.multi
+	def open_line_current(self):
+		"""
+		Open line current
+		"""
+		res_id = self.id
+		return {
+				'type': 'ir.actions.act_window',
+				'name': ' Edit Order Current',
+				'view_type': 'form',
+				'view_mode': 'form',
+				'res_model': self._name,
+				'res_id': res_id,
+				'target': 'current',
+				'flags': {
+						'form': {'action_buttons': True, }
+						},
+				'context': {}
+		}
+	# open_line_current
+
+# End ----------------------------------- Inherited from mgt_line.py ------------------------------
