+++ models/management/mgt_family_line.py
@@ -3,7 +3,7 @@
  	Family Line 
 
 	Created: 			20 Aug 2018
-	Last up: 			18 Dec 2020
+	Last up: 			24 mar 2021
 """
 from __future__ import print_function
 from openerp import models, fields, api
@@ -20,7 +20,7 @@ class MgtFamilyLine(models.Model):
 	"""
 	_name = 'openhealth.management.family.line'
 	_order = 'amount desc'
-	_inherit = 'openhealth.management.line'
+	#_inherit = 'openhealth.management.line'
 
 # ----------------------------------------------------------- Update -----------
 	def update(self):
@@ -81,3 +81,78 @@ class MgtFamilyLine(models.Model):
 			self.meta_sp = 'Otros'
 
 	# update_fields
+
+
+# Begin ----------------------------------- Inherited from mgt_line.py ----------------------------
+
+# ----------------------------------------------------------- Interface --------
+	management_id = fields.Many2one(
+			'openhealth.management',
+		)
+
+	doctor_id = fields.Many2one(
+			'openhealth.management.doctor.line',
+			ondelete='cascade',
+		)
+
+# ----------------------------------------------------------- Primitive --------
+	name = fields.Char(
+			'Name',
+		)
+
+	name_sp = fields.Char(
+			'Nombre',
+		)
+
+	meta = fields.Char(
+			'Meta',
+		)
+
+	meta_sp = fields.Char(
+			'Meta',
+		)
+
+	idx = fields.Integer(
+			'Idx',
+		)
+
+	x_count = fields.Integer(
+			'Nr',
+		)
+
+	amount = fields.Float(
+			'Monto',
+			digits=(16, 1),
+		)
+
+	per_amo = fields.Float(
+			'% Monto',
+		)
+
+	per_nr = fields.Float(
+			'% Nr',
+		)
+
+#----------------------------------------------------------- Method ------------
+	@api.multi
+	def open_line_current(self):
+		"""
+		Open line current
+		"""
+		res_id = self.id
+		return {
+				'type': 'ir.actions.act_window',
+				'name': ' Edit Order Current',
+				'view_type': 'form',
+				'view_mode': 'form',
+				'res_model': self._name,
+				'res_id': res_id,
+				'target': 'current',
+				'flags': {
+						'form': {'action_buttons': True, }
+						},
+				'context': {}
+		}
+	# open_line_current
+
+# End ----------------------------------- Inherited from mgt_line.py ------------------------------
