+++ models/management/mgt_subfamily_line.py
@@ -2,10 +2,8 @@
 """
 	Sub Family Line 
 
-	Only Data model. No functions.
-
 	Created: 			20 Aug 2018
-	Last up: 			18 Dec 2020
+	Last up: 			24 mar 2021
 """
 from __future__ import print_function
 from openerp import models, fields, api
@@ -137,7 +135,7 @@ class MgtSubfamilyLine(models.Model):
 			self.meta_sp = 'Otros'
 	# update
 
-#-------------------------------------------------------- Open Line Current ----
+#----------------------------------------------------------- Method ------------
 	@api.multi
 	def open_line_current(self):
 		"""
@@ -153,7 +151,6 @@ class MgtSubfamilyLine(models.Model):
 				'res_id': res_id,
 				'target': 'current',
 				'flags': {
-						#'form': {'action_buttons': True, 'options': {'mode': 'edit'}}
 						'form': {'action_buttons': True, }
 						},
 				'context': {}
