+++ models/management_doctor/__init__.py
@@ -6,10 +6,8 @@
 # - Isolate 
 # - Reduce inheritance
 
-#from . import day_doctor_line   # For doctor_line
 from . import mgt_day_doctor_line   # For doctor_line
 
-#from . import doctor_line      # Too complex 
 from . import mgt_doctor_line      # Too complex 
 
 from . import doctor_daily
