+++ models/management_doctor/mgt_doctor_line.py
@@ -1,11 +1,8 @@
 # -*- coding: utf-8 -*-
 """
-	Doctor Line - Too complex - Refactor !
+	Doctor Line
 
-	Inherits  - mgt_line
-
-	Why is this too complex ?
-	
+	Too complex - Refactor !
 		- It has too much crossed relations:  5 One2many (has many)
 		- It has a confusing inheritance. Is this necessary ?
 
@@ -14,26 +11,25 @@
 		- Refactor into a completely independent new model. Follow Marketing approach. 
 
 	Created: 			18 may 2018
-	Last up: 			19 oct 2020
+	Last up: 			24 mar 2021
 """
 from __future__ import print_function
 import collections
 from openerp import models, fields, api
 
-#class DoctorLine(models.Model):
 class MgtDoctorLine(models.Model):
 	"""
     Mgt Doctor line
 	"""
 	_name = 'openhealth.management.doctor.line'
 	_order = 'amount desc'
-	_inherit = 'openhealth.management.line'
+	#_inherit = 'openhealth.management.line'
 
 
 # ----------------------------------------------------------- Relations --------
-	management_id = fields.Many2one(
-			'openhealth.management',
-		)
+	#management_id = fields.Many2one(
+	#		'openhealth.management',
+	#	)
 
 
 # ----------------------------------------------------------- One2many ---------
@@ -329,3 +325,79 @@ class MgtDoctorLine(models.Model):
 		#self.update() # Generates error !
 
 	# stats
+
+
+
+# Begin ----------------------------------- Inherited from mgt_line.py ----------------------------
+
+# ----------------------------------------------------------- Interface --------
+	management_id = fields.Many2one(
+			'openhealth.management',
+		)
+
+	doctor_id = fields.Many2one(
+			'openhealth.management.doctor.line',
+			ondelete='cascade',
+		)
+
+# ----------------------------------------------------------- Primitive --------
+	name = fields.Char(
+			'Name',
+		)
+
+	name_sp = fields.Char(
+			'Nombre',
+		)
+
+	meta = fields.Char(
+			'Meta',
+		)
+
+	meta_sp = fields.Char(
+			'Meta',
+		)
+
+	idx = fields.Integer(
+			'Idx',
+		)
+
+	x_count = fields.Integer(
+			'Nr',
+		)
+
+	amount = fields.Float(
+			'Monto',
+			digits=(16, 1),
+		)
+
+	per_amo = fields.Float(
+			'% Monto',
+		)
+
+	per_nr = fields.Float(
+			'% Nr',
+		)
+
+#----------------------------------------------------------- Method ------------
+	@api.multi
+	def open_line_current(self):
+		"""
+		Open line current
+		"""
+		res_id = self.id
+		return {
+				'type': 'ir.actions.act_window',
+				'name': ' Edit Order Current',
+				'view_type': 'form',
+				'view_mode': 'form',
+				'res_model': self._name,
+				'res_id': res_id,
+				'target': 'current',
+				'flags': {
+						'form': {'action_buttons': True, }
+						},
+				'context': {}
+		}
+	# open_line_current
+
+# End ----------------------------------- Inherited from mgt_line.py ------------------------------
